# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * Life Simulation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SimulationState {
  tick: number;
  running: boolean;
  gameHour: number;
  gameDay: number;
  population: number;
  avgMood: number;
  gdp: number;
  unemploymentRate: number;
  governmentBudget: number;
  totalTaxCollected: number;
  avgWealth: number;
}

export interface Agent {
  id: number;
  name: string;
  gender: string;
  age: number;
  mood: number;
  money: number;
  personality: string;
  socialization: number;
  currentAction: string;
  /** @nullable */
  employerId?: number | null;
  isRetired?: boolean;
  careerLevel?: number;
  jobTitle?: string;
}

export interface AgentNeeds {
  hunger: number;
  comfort: number;
  social: number;
  health: number;
  sleep: number;
  education?: number;
  entertainment?: number;
  faith?: number;
  housingSafety?: number;
  financialSafety?: number;
  physicalSafety?: number;
  socialRating?: number;
  wellbeing?: number;
}

export interface AgentRelation {
  otherId: number;
  otherName: string;
  friendshipLevel: number;
}

export interface JobHistoryEntry {
  tick: number;
  event: 'hired' | 'fired' | 'retired' | 'quit' | 'promoted';
  /** @nullable */
  businessId: number | null;
  /** @nullable */
  businessName: string | null;
  duration?: number;
}

export type AgentDetail = Agent & {
  needs: AgentNeeds;
  relations: AgentRelation[];
  /** Last 10 actions taken by this agent */
  recentActions: string[];
  /** Whether this agent has retired */
  isRetired: boolean;
  /** Most recent 20 job events for this agent */
  jobHistory: JobHistoryEntry[];
  /** Name of current employer */
  employerName?: string | null;
  /** Tick when current job started */
  jobStartTick?: number | null;
  /** Ticks at current job */
  jobTenure?: number | null;
  /** Total number of jobs held */
  totalJobs?: number;
  /** Number of promotions received */
  promotions?: number;
  /** Current career grade 1-5 */
  careerLevel?: number;
  /** Ambition attribute 20-100 */
  ambition?: number;
  /** Target career grade derived from ambition */
  targetCareerLevel?: number;
  /** Physical strength attribute 1-100 */
  strength?: number;
  /** Intelligence attribute 1-100 */
  intelligence?: number;
};

export interface AgentListResponse {
  agents: Agent[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface Business {
  id: number;
  name: string;
  type: string;
  balance: number;
  productionRate: number;
  employeeCount: number;
  /** @nullable */
  ownerId?: number | null;
  firedThisTick: number;
  hiredThisTick: number;
  productivityLevel: number;
}

export interface Good {
  id: number;
  name: string;
  basePrice: number;
  currentPrice: number;
  quality: number;
  demand: number;
  supply: number;
}

export interface GovernmentState {
  budget: number;
  totalTaxCollected: number;
  totalSubsidiesPaid: number;
  totalPensionPaid: number;
  totalPublicServicesPaid: number;
  taxRate: number;
  subsidyAmount: number;
  pensionRate: number;
  totalGrantsPaid: number;
  grantsIssuedLastDay: number;
  unemploymentRatePct: number;
  grantThresholdPct: number;
}

export interface SimConfig {
  /** Tax rate as a fraction (0.0 - 1.0) */
  taxRate: number;
  /** How fast needs decay per tick (0.0 - 100.0) */
  needDecayRate: number;
  /** Milliseconds per simulation tick */
  tickIntervalMs: number;
  /** Number of agents on reset */
  initialAgents: number;
  /** Number of businesses on reset */
  initialBusinesses: number;
  /** Base price for food goods */
  baseFoodPrice: number;
  /** Base salary per tick for employed agents */
  baseSalary: number;
  /** Amount paid to agents with no money */
  subsidyAmount: number;
  /** How much social interactions affect mood (0.0-10.0) */
  socialInteractionStrength: number;
  /** Business markup on top of base price (0.0 - 1.0) */
  priceMarkup: number;
  /** Pension as a fraction of baseSalary paid to retirees per tick (0.0 - 1.0) */
  pensionRate: number;
}

export interface SimConfigUpdate {
  taxRate?: number;
  needDecayRate?: number;
  tickIntervalMs?: number;
  initialAgents?: number;
  initialBusinesses?: number;
  baseFoodPrice?: number;
  baseSalary?: number;
  subsidyAmount?: number;
  socialInteractionStrength?: number;
  priceMarkup?: number;
  pensionRate?: number;
}

export interface StatsSnapshot {
  tick: number;
  gameHour: number;
  gameDay: number;
  avgMood: number;
  gdp: number;
  population: number;
  avgWealth: number;
  unemploymentRate: number;
  governmentBudget: number;
  recordedAt: string;
}

export interface StatsSummary {
  totalAgents: number;
  totalBusinesses: number;
  totalGoods: number;
  employedAgents: number;
  unemployedAgents: number;
  avgMood: number;
  avgWealth: number;
  avgHealth: number;
  avgSleep: number;
  richestAgent?: string;
  happiestAgent?: string;
  mostPopularGood?: string;
  gdp: number;
  birthsLastTick: number;
  deathsLastTick: number;
  profitableBusinesses: number;
  unprofitableBusinesses: number;
  marketBalance: number;
  totalDemand: number;
  totalSupply: number;
}

export interface TopAgentsResponse {
  byWealth: Agent[];
  byMood: Agent[];
  byAge: Agent[];
  bySocialization: Agent[];
}

export interface ErrorResponse {
  error: string;
}

export type ListAgentsParams = {
page?: number;
limit?: number;
sortBy?: ListAgentsSortBy;
sortDir?: ListAgentsSortDir;
filterAction?: string;
};

export type ListAgentsSortBy = typeof ListAgentsSortBy[keyof typeof ListAgentsSortBy];


export const ListAgentsSortBy = {
  name: 'name',
  age: 'age',
  mood: 'mood',
  money: 'money',
  currentAction: 'currentAction',
} as const;

export type ListAgentsSortDir = typeof ListAgentsSortDir[keyof typeof ListAgentsSortDir];


export const ListAgentsSortDir = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type GetStatsHistoryParams = {
limit?: number;
};

export type GetTopAgentsParams = {
limit?: number;
};