# lib/api-zod/src/generated/types/simConfigUpdate.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * Life Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface SimConfigUpdate {
  taxRate?: number;
  needDecayRate?: number;
  tickIntervalMs?: number;
  initialAgents?: number;
  initialBusinesses?: number;
  baseFoodPrice?: number;
  baseSalary?: number;
  subsidyAmount?: number;
  socialInteractionStrength?: number;
  priceMarkup?: number;
  pensionRate?: number;
}